Attempting to give a better error message for function calls with spaces in them, i.e. `f ()`, `g (x)`, `h (y, z)`, etc.
This is *hacky* and not very robust: we detect unary cases as well-formedness errors, and everything else as "parse errors".  The well-formedness error has to attempt to unwrap the pieces it receives, to undo the builtin["trace-value"] wrapping.  The "parse errors" actually successfully parse, but complain anyway!  The grammar rules for handling 0-arity and 2+arity are carefully rigged to be disjoint from the unary case, so that we avoid ambiguous parses; I'm only *mostly* confident that they really are unambiguous.